+++ src/App.js
@@ -2,7 +2,6 @@ import React from 'react'
 import { Provider } from 'react-redux'
 import { combineReducers, configureStore } from '@reduxjs/toolkit'
 import { quiz } from 'reducers/quiz'
-import { useSelector } from 'react-redux'
 
 
 import { CurrentQuestion } from 'components/CurrentQuestion'
